+++ ghost/core/core/frontend/web/site.js
@@ -14,7 +14,7 @@ const themeEngine = require('../services/theme-engine');
 const themeMiddleware = themeEngine.middleware;
 const membersService = require('../../server/services/members');
 const offersService = require('../../server/services/offers');
-const customRedirects = require('../../server/services/redirects');
+const customRedirects = require('../../server/services/custom-redirects');
 const siteRoutes = require('./routes');
 const shared = require('../../server/web/shared');
 const errorHandler = require('@tryghost/mw-error-handler');
