+++ ghost/core/core/server/api/endpoints/redirects.js
@@ -1,6 +1,6 @@
 const path = require('path');
 
-const redirects = require('../../services/redirects');
+const customRedirects = require('../../services/custom-redirects');
 
 module.exports = {
     docName: 'redirects',
@@ -10,7 +10,7 @@ module.exports = {
             disposition: {
                 type: 'file',
                 value() {
-                    return redirects.api.getRedirectsFilePath()
+                    return customRedirects.api.getRedirectsFilePath()
                         .then((filePath) => {
                             // @deprecated: .json was deprecated in v4.0 but is still the default for backwards compat
                             return filePath === null || path.extname(filePath) === '.json'
@@ -23,13 +23,13 @@ module.exports = {
         permissions: true,
         response: {
             async format() {
-                const filePath = await redirects.api.getRedirectsFilePath();
+                const filePath = await customRedirects.api.getRedirectsFilePath();
 
                 return filePath === null || path.extname(filePath) === '.json' ? 'json' : 'plain';
             }
         },
         query() {
-            return redirects.api.get();
+            return customRedirects.api.get();
         }
     },
 
@@ -39,7 +39,7 @@ module.exports = {
             cacheInvalidate: true
         },
         query(frame) {
-            return redirects.api.setFromFilePath(frame.file.path, frame.file.ext);
+            return customRedirects.api.setFromFilePath(frame.file.path, frame.file.ext);
         }
     }
 };
